+++ crates/ink/tests/ui/abi/sol/pass/message-selector-encoding-args.rs
@@ -20,7 +20,7 @@ mod contract {
         pub fn message_1(&self, _input_1: u8) {}
 
         #[ink(message, selector = 0xC0DE_CAFE)]
-        pub fn message_2(&self, _input_1: SolBytes<[u8; 32]>) {}
+        pub fn message_2(&self, _input_1: ink::sol::FixedBytes<32>) {}
 
         #[ink(message)]
         pub fn message_3(&self, _input_1: bool, _input_2: i8) {}
@@ -29,12 +29,12 @@ mod contract {
     #[ink::trait_definition]
     pub trait Messages {
         #[ink(message, selector = 0x12345678)]
-        fn message_4(&self, _input_1: SolBytes<Vec<u8>>);
+        fn message_4(&self, _input_1: ink::sol::DynBytes);
     }
 
     impl Messages for Contract {
         #[ink(message, selector = 0x12345678)]
-        fn message_4(&self, _input_1: SolBytes<Vec<u8>>) {}
+        fn message_4(&self, _input_1: ink::sol::DynBytes) {}
     }
 
     impl Contract {
@@ -46,8 +46,8 @@ mod contract {
             &self,
             _input_1: bool,
             _input_2: String,
-            _input_3: SolBytes<[u8; 32]>,
-            _input_4: SolBytes<Vec<u8>>,
+            _input_3: ink::sol::FixedBytes<32>,
+            _input_4: ink::sol::DynBytes,
             _input_5: [u8; 4],
             _input_6: Vec<u8>,
         ) {